app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
celery[redis]==5.4.0
redis==5.3.1
httpx==0.28.0
python-multipart==0.0.17
alembic==1.14.0
faster-whisper==1.1.0
requests==2.32.3
